remove syscallctx; minor bugs fixed

+++ anemone-kernel/src/arch/loongarch64/exception/trap/mod.rs
@@ -111,9 +111,9 @@ impl TrapFrameArch for LA64TrapFrame {
     };
 
     /// Read syscall argument IDX from the trap frame.
-    unsafe fn syscall_args<const IDX: usize>(&self) -> usize {
+    unsafe fn syscall_args<const IDX: usize>(&self) -> u64 {
         const_assert!(IDX < 7);
-        self.gpr.a::<IDX>() as usize
+        self.gpr.a::<IDX>()
     }
 
     unsafe fn syscall_no(&self) -> usize {
